`size_t` in `mojo::DataPipe[Consumer|Producer]Handle`: /storage.
This CL makes use of the new `size_t`-based APIs in the remaining
directories like `//components`, `//extensions`, or `//storage`.
See https://crrev.com/c/5414148 for motivation.

This CL may have introduced new integer conversions - in those cases the
CL attempts to have well-defined behavior (e.g. using
`base::checked_cast`).  OTOH, this CL did *not* attempt to fix old
integer operations (e.g. old `int` => `uint32_t` / `size_t` conversions,
or old additions that may potentially overflow).

This CL was uploaded by git cl split.

R=asully@chromium.org, mek@chromium.org

Bug: 40284755
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5461384
Commit-Queue: Austin Sullivan <asully@chromium.org>
Auto-Submit: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Łukasz Anforowicz <lukasza@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1288800}

+++ storage/browser/file_system/file_writer_delegate.cc
@@ -13,6 +13,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/functional/bind.h"
 #include "base/functional/callback.h"
 #include "base/location.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/task/sequenced_task_runner.h"
 #include "base/task/single_thread_task_runner.h"
 #include "base/threading/thread_restrictions.h"
@@ -148,7 +149,7 @@ void FileWriterDelegate::Read() {
   }
 
   DCHECK(data_pipe_);
-  uint32_t num_bytes = io_buffer_->size();
+  size_t num_bytes = io_buffer_->size();
   MojoResult result = data_pipe_->ReadData(io_buffer_->data(), &num_bytes,
                                            MOJO_READ_DATA_FLAG_NONE);
   if (result == MOJO_RESULT_SHOULD_WAIT) {
@@ -156,7 +157,7 @@ void FileWriterDelegate::Read() {
     return;
   }
   if (result == MOJO_RESULT_OK) {
-    bytes_read_ = num_bytes;
+    bytes_read_ = base::checked_cast<int>(num_bytes);
     OnReadCompleted(bytes_read_);
     return;
   }
